fix emulator bug
