+++ legacy/firmware/trezor.c
@@ -144,6 +144,9 @@ int main(void) {
 #endif
 #endif
 
+#if EMULATOR
+  g_ucWorkMode = WORK_MODE_USB;
+#endif
   vlayoutLogo();
   oledRefresh();
